front: a local takes the type Go gave it, and int and float64 are refused there
A local was typed from its initializer's IR type, and constType maps an
untyped int, int64, int16 and float64 constant to i32 or f32, so `n := 0`
and `x := 1.5` were lowered as int32 and float32 while the authored
function ran them as int and float64; `n := len(out)` did the same through
Len's i32. The platform-width refusal irType gives `n + 1` never applied to
`n++` or to the declaration itself.

A local's Go type is checked at the declaration: a basic type has to pass
irType's rule, and float64 is named for what it is rather than as a
binding element. The corpus's Pack counted an axis down with an int loop
variable and is spelled int32 now; its source digest moves, and the
generated file and the emit golden carry it.

The four rejection cases -- a local, a loop variable, a float64 local and
a len-typed local -- were accepted before this change.

+++ internal/testkernels/pack.go
@@ -64,7 +64,7 @@ func Pack(t accel.Thread, p PackParams, src []float32, dst []float32) {
 	if i < p.Count {
 		rem := i
 		at := p.Offset
-		for axis := PackRank - 1; axis >= 0; axis-- {
+		for axis := int32(PackRank - 1); axis >= 0; axis-- {
 			a := uint32(axis)
 			if a < p.Rank {
 				e := p.Extent[a]
